Add full functionality from develop
* init: create new branch develop

* feat(films): add directors

* Добавил вывод общих фильмов с сортировкой по популярности

* в ожидании доступа к режиссеру

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Реализована фича "Отзывы"

* Removing movies and users.

* "Recommendations" functionality

* "Recommendations" functionality

* "Recommendations" functionality

* "Recommendations" functionality

* fix: устранена ошибка при добавлении нового фильма в FilmServiceImpl

* Добавил вывод общих фильмов с сортировкой по популярности

* в ожидании доступа к режиссеру

* Добавил поиск по названию и режиссеру

* Добавил поиск по названию и режиссеру

* Функциональность "Отзывы" (#7)

* Реализована фича "Отзывы"

* fix

* fix

* fix Checkstyle

* Checkstyle

* Checkstyle fix

* исправления для влития в ветку

* исправления для влития в ветку 2

* исправления для влития в ветку

* feat: add features for Filmorate project

* fix: set name to user if not exists

* Исправлены ошибки в сотртировке отзывов

* Исправлены ошибки в сотртировке отзывов

* Исправлены ошибки в сотртировке отзывов

* fix: numbering mpa and genres

* fix: update film operation

* fix: update film operation

* fix: update film operation

* Revert "Add reviews"

* fix

* fix: some bugs

* fix: some bugs

* fix: some bugs

* fix: some bugs

* fix: some bugs

* CheckStyle

* CheckStyle

* Исправлены ошибки в сотртировке отзывов (#15)

* Исправил add-feed

* Исправил add-feed

* Исправил add-feed

---------

+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/FilmServiceImpl.java
@@ -3,12 +3,16 @@ package ru.yandex.practicum.filmorate.service.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.mappers.FilmMapper;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.mappers.filmMap.FilmMapper;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.FilmSortBy;
+import ru.yandex.practicum.filmorate.model.UserEvents;
 import ru.yandex.practicum.filmorate.repository.FilmRepository;
-import ru.yandex.practicum.filmorate.repository.dto.FilmDto;
-import ru.yandex.practicum.filmorate.repository.dto.NewFilmRequest;
-import ru.yandex.practicum.filmorate.repository.dto.UpdateFilmRequest;
+import ru.yandex.practicum.filmorate.repository.UserRepository;
+import ru.yandex.practicum.filmorate.repository.dto.film.FilmDto;
+import ru.yandex.practicum.filmorate.repository.dto.film.NewFilmRequest;
+import ru.yandex.practicum.filmorate.repository.dto.film.UpdateFilmRequest;
 import ru.yandex.practicum.filmorate.service.FilmService;
 import ru.yandex.practicum.filmorate.util.FilmValidator;
 
@@ -23,13 +27,15 @@ import java.util.stream.Collectors;
 public class FilmServiceImpl implements FilmService {
 
     private final FilmRepository filmRepository;
+    private final UserRepository userRepository;
+    private final FeedEventService feedEventService;
 
     @Override
     public FilmDto addFilm(NewFilmRequest newFilmRequest) {
         FilmValidator.validator(newFilmRequest);
         Film film = filmRepository.addFilm(FilmMapper.mapToFilm(newFilmRequest));
         log.info("FilmServiceImpl: новый фильм {}, с id {} добавлен", film.getName(), film.getId());
-        return FilmMapper.mapToFilmDto(filmRepository.addFilm(film));
+        return FilmMapper.mapToFilmDto(film);
     }
 
     @Override
@@ -59,22 +65,65 @@ public class FilmServiceImpl implements FilmService {
     @Override
     public void addLike(int filmId, int userId) {
         filmRepository.addLike(filmId, userId);
+        feedEventService.addEvent(userId, filmId, UserEvents.EventType.LIKE, UserEvents.Operation.ADD);
         log.info("FilmServiceImpl: Пользователю с id {} нравится фильм с id: {}", userId, filmId);
     }
 
     @Override
     public void removeLike(int filmId, int userId) {
+        if (!filmRepository.filmExists(filmId)) {
+            throw new NotFoundException("пользователь с id: " + userId + " не найден");
+        }
+        if (!userRepository.userExists(userId)) {
+            throw new NotFoundException("пользователь с id: " + userId + " не найден");
+        }
         filmRepository.removeLike(filmId, userId);
+        feedEventService.addEvent(userId, filmId, UserEvents.EventType.LIKE, UserEvents.Operation.REMOVE);
         log.info("FilmServiceImpl: Пользователю с id {} перестал нравится фильм с id: {}", userId, filmId);
     }
 
     @Override
-    public Set<FilmDto> getTopFilms(int count) {
-        log.info("FilmServiceImpl: запрос топ-" + count + " фильмов");
-        List<Film> topPopular = filmRepository.getTopPopular(count);
+    public Set<FilmDto> getTopFilms(int count, Integer genreId, Integer year) {
+        log.info("FilmServiceImpl: запрос топ-" + count + " фильмов с параметрами (" + genreId + "," + year + " )");
+        List<Film> topPopular = filmRepository.getTopPopular(count, genreId, year);
         return topPopular.stream()
                 .map(FilmMapper::mapToFilmDto)
                 .collect(Collectors.toCollection(LinkedHashSet::new));
+    }
+
+    @Override
+    public List<FilmDto> getCommonFilms(int userId, int friendId) {
+        log.debug("FilmServiceImpl: запрос в сервис: userId={}, friendId={}", userId, friendId);
+        List<Film> commonFilms = filmRepository.getCommonFilms(userId, friendId);
+        return commonFilms.stream()
+                .map(FilmMapper::mapToFilmDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<FilmDto> searchFilms(String query, String by) {
+        List<Film> searchFilms = filmRepository.search(query, by);
+        return searchFilms.stream()
+                .map(FilmMapper::mapToFilmDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<FilmDto> getFilmsByDirector(int directorId, FilmSortBy sortBy) {
+        log.info("FilmController: запрошен список фильмов режиссера с id = {} и отсортированный по {}", directorId, sortBy);
+        List<Film> filmByDirector = filmRepository.getFilmsByDirector(directorId, sortBy);
+        return filmByDirector.stream()
+                .map(FilmMapper::mapToFilmDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteFilm(int filmId) {
+        if (!filmRepository.filmExists(filmId)) {
+            throw new NotFoundException("FilmServiceImpl: фильм с id: " + filmId + " не найден");
+        }
+        filmRepository.deleteFilm(filmId);
+        log.info("FilmServiceImpl: фильм с id: {} удалён", filmId);
     }
 
 }
